// Frontend/Angular/src/app/clients/student/model/lessonDto.ts
/**
 * Student api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AttachmentDto } from './attachmentDto';

export interface LessonDto { 
    id?: number;
    name?: string;
    description?: string;
    order?: number;
    isPractical?: boolean;
    subjectId?: number;
    subjectName?: string;
    assignmentId?: number;
    attachments?: Array<AttachmentDto>;
}